debug issues and run some manual tests
